Hack around an 3D simulation  bug that needs a more permanent fix in the near future.
Namely HRTF wouldn't enable when using environments.  With the new property callback mechanism, we can't have out-of-line nodes.

The fix for this and other things is a visitor as opposed to recursive tick functions, but this has not been implemented as of yet, and work on other branches is more pressing.